Updated code according to clang warnings and suggestions

+++ common.h
@@ -10,6 +10,6 @@
 
 constexpr double pi = 3.1415926535897932384626433832795028841971693;
 constexpr double pi2 = pi * 2;
-constexpr float pif = pi;
-constexpr float pi2f = pi2;
+constexpr float pif = static_cast<float>(pi);
+constexpr float pi2f = static_cast<float>(pi2);
